Modified PCB for 10ch PDM board
I modified the original dingoPDM for 10 channels. The board has been manufactured, tested and works but after mounting I decided that it doesn't suit my needs for the current build. I may develop it further in the future, make it pretty etc but for now I am marking this project as stale.